backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain
langchain-google-genai
google-generativeai>=0.8.0
daytona>=0.1.0
